\name{print.howmany}
\alias{print.howmany}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Print an object of class 'howmany' }
\description{
    Displays summary information about an object of class \code{howmany}
}
\usage{
\method{print}{howmany}(x,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ an object of class \code{howmany} }
  \item{...}{ additional arguments, currently un-implemented }
}
\details{
  Currently, invokes the \code{summary} method
}
\seealso{ \code{\link{howmany}} and \code{\link{howmany_dependent}};
  \code{\link{summary.howmany}} and \code{\link{plot.howmany}} }

\keyword{htest}

