\name{summary.howmany}
\alias{summary.howmany}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Summary of an object of class 'howmany' }
\description{
  Displays summary information about an object of class \code{howmany}
}
\usage{
\method{summary}{howmany}(object,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{ an object of class \code{howmany} }
  \item{...}{ additional arguments, currently un-implemented }
}

\seealso{ \code{\link{howmany}} and \code{\link{howmany_dependent}};
  \code{\link{plot.howmany}} }

\keyword{htest}
